xen/x86: manually build xen.mb.efi binary
This patch introduces xen.mb.efi which contains PE header build
manually.
The code executed by it is almost the same like the code executed by
currently existing xen.efi. However, the file layout is simpler and
completely different. This way we have better control on PE image.
Hence,
it allow us to play tricks which are not feasible in xen.efi.

This is the first step to get:
  - one binary which can be loaded by the UEFI loader, Multiboot and
    Multiboot2 protocols,
  - UEFI secure boot support when Xen is loaded via Multiboot2 protocol,
  - if we wish, in the future we can drop xen/xen.gz and build
    xen.mb.efi only,
  - crash dumps generated by the xen.mb.efi loaded from the EFI loader
    can be analyzed by crash tool,
  - simpler code,
  - simpler build,
  - Xen build will no longer depend on ld i386pep support.

Signed-off-by: Bobby Eshleman <bobbyeshleman@gmail.com>